breaking: removed all supabase classes and consolidate all auth related logic to SupabaseAuth
