--- lib/src/supabase_deep_linking_mixin.dart
@@ -1,72 +1,0 @@
-import 'dart:async';
-
-import 'package:flutter/foundation.dart';
-import 'package:flutter/services.dart';
-import 'package:flutter/widgets.dart';
-
-import 'package:supabase_flutter/supabase_flutter.dart';
-import 'package:uni_links/uni_links.dart';
-
-mixin SupabaseDeepLinkingMixin<T extends StatefulWidget> on State<T> {
-  StreamSubscription<Uri?>? _sub;
-
-  void startDeeplinkObserver() {
-    Supabase.instance.log('***** SupabaseDeepLinkingMixin startAuthObserver');
-    _handleIncomingLinks();
-    _handleInitialUri();
-  }
-
-  void stopDeeplinkObserver() {
-    Supabase.instance.log('***** SupabaseDeepLinkingMixin stopAuthObserver');
-    if (_sub != null) _sub?.cancel();
-  }
-
-  /// Handle incoming links - the ones that the app will recieve from the OS
-  /// while already started.
-  void _handleIncomingLinks() {
-    if (!kIsWeb) {
-      // It will handle app links while the app is already started - be it in
-      // the foreground or in the background.
-      _sub = uriLinkStream.listen(
-        (Uri? uri) {
-          if (mounted && uri != null) {
-            handleDeeplink(uri);
-          }
-        },
-        onError: (Object err) {
-          if (!mounted) return;
-          onErrorReceivingDeeplink(err.toString());
-        },
-      );
-    }
-  }
-
-  /// Handle the initial Uri - the one the app was started with
-  ///
-  /// **ATTENTION**: `getInitialLink`/`getInitialUri` should be handled
-  /// ONLY ONCE in your app's lifetime, since it is not meant to change
-  /// throughout your app's life.
-  ///
-  /// We handle all exceptions, since it is called from initState.
-  Future<void> _handleInitialUri() async {
-    if (!SupabaseAuth.instance.shouldHandleInitialDeeplink()) return;
-
-    try {
-      final uri = await getInitialUri();
-      if (mounted && uri != null) {
-        handleDeeplink(uri);
-      }
-    } on PlatformException {
-      // Platform messages may fail but we ignore the exception
-    } on FormatException catch (err) {
-      if (!mounted) return;
-      onErrorReceivingDeeplink(err.message);
-    }
-  }
-
-  /// Callback when deeplink receiving succeeds
-  void handleDeeplink(Uri uri);
-
-  /// Callback when deeplink receiving throw error
-  void onErrorReceivingDeeplink(String message);
-}
